multiGraphDomainClipGeometry param removed from the TravelOptions to use only clipGeometry param for clipping.
